+++ tests/TestCase/Controller/PostsCategoriesControllerTest.php
@@ -1,0 +1,153 @@
+<?php
+/**
+ * This file is part of MeCms.
+ *
+ * MeCms is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * MeCms is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with MeCms.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ */
+namespace MeCms\Test\TestCase\Controller;
+
+use Cake\Cache\Cache;
+use Cake\ORM\TableRegistry;
+use Cake\TestSuite\IntegrationTestCase;
+
+/**
+ * PostsCategoriesControllerTest class
+ */
+class PostsCategoriesControllerTest extends IntegrationTestCase
+{
+    /**
+     * @var \MeCms\Model\Table\PostsCategoriesTable
+     */
+    protected $PostsCategories;
+
+    /**
+     * Fixtures
+     * @var array
+     */
+    public $fixtures = [
+        'plugin.me_cms.posts',
+        'plugin.me_cms.posts_categories',
+        'plugin.me_cms.posts_tags',
+        'plugin.me_cms.tags',
+        'plugin.me_cms.users',
+    ];
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->PostsCategories = TableRegistry::get('MeCms.PostsCategories');
+
+        Cache::clear(false, $this->PostsCategories->cache);
+    }
+
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        unset($this->PostsCategories);
+    }
+
+    /**
+     * Adds additional event spies to the controller/view event manager
+     * @param \Cake\Event\Event $event A dispatcher event
+     * @param \Cake\Controller\Controller|null $controller Controller instance
+     * @return void
+     */
+    public function controllerSpy($event, $controller = null)
+    {
+        $controller->viewBuilder()->setLayout(false);
+
+        parent::controllerSpy($event, $controller);
+    }
+
+    /**
+     * Tests for `index()` method
+     * @test
+     */
+    public function testIndex()
+    {
+        $this->get(['_name' => 'postsCategories']);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/PostsCategories/index.ctp');
+
+        $categoriesFromView = $this->viewVariable('categories');
+        $this->assertInstanceof('Cake\ORM\ResultSet', $categoriesFromView);
+
+        foreach ($categoriesFromView as $category) {
+            $this->assertInstanceof('MeCms\Model\Entity\PostsCategory', $category);
+        }
+
+        $cache = Cache::read('categories_index', $this->PostsCategories->cache);
+        $this->assertEquals($categoriesFromView->toArray(), $cache->toArray());
+    }
+
+    /**
+     * Tests for `view()` method
+     * @test
+     */
+    public function testView()
+    {
+        $slug = $this->PostsCategories->find('active')
+            ->order([sprintf('%s.id', $this->PostsCategories->getAlias()) => 'ASC'])
+            ->extract('slug')
+            ->first();
+
+        $url = ['_name' => 'postsCategory', $slug];
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/PostsCategories/view.ctp');
+
+        $categoryFromView = $this->viewVariable('category');
+        $this->assertInstanceof('MeCms\Model\Entity\PostsCategory', $categoryFromView);
+
+        $postsFromView = $this->viewVariable('posts');
+        $this->assertInstanceof('Cake\ORM\ResultSet', $postsFromView);
+
+        foreach ($postsFromView as $post) {
+            $this->assertInstanceof('MeCms\Model\Entity\Post', $post);
+        }
+
+        //Sets the cache name
+        $cache = sprintf('category_%s_limit_%s_page_%s', md5($slug), config('default.records'), 1);
+        list($postsFromCache, $pagingFromCache) = array_values(Cache::readMany(
+            [$cache, sprintf('%s_paging', $cache)],
+            $this->PostsCategories->cache
+        ));
+
+        $this->assertEquals($postsFromView->toArray(), $postsFromCache->toArray());
+        $this->assertNotEmpty($pagingFromCache['Posts']);
+
+        $this->get(array_merge($url, ['?' => ['q' => $slug]]));
+        $this->assertRedirect($url);
+    }
+}