+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.17 branch
+### 2.17.4
+* added tests for `PostsCategoriesController` class.
+
 ### 2.17.3
 * added `PhotosTable::findPending()` method;
 * fixed all `find()` methods;